fix: ensure use directives execute in the correct sequence
